pyproject.toml
# pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "policyengine_tests_generator"
version = "0.1.3"
authors = [
  { name = "Al Noman", email = "al@policyengine.org" },
]
description = "A test case YAML generator for PolicyEngine simulations"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyyaml>=6.0",
    "typing-extensions>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["policyengine_tests_generator"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "isort>=5.0",
    "mypy>=0.900",
    "numpy",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
pythonpath = [
  "."
]